Restructured rendering module
