+++ Engine/Source/Application.cpp
@@ -8,9 +8,7 @@
 #include "ModuleCamera.h"
 #include "ModuleDebugDraw.h"
 #include "ModuleFileSystem.h"
-#include "ModuleProgram.h"
 #include "ModuleTimer.h"
-#include "ModuleSkybox.h"
 
 
 Application::Application()
@@ -18,15 +16,13 @@ Application::Application()
 	// Order matters: they will Init/start/update in this order
 	modules[0] = input = new ModuleInput();
 	modules[1] = window = new ModuleWindow();
-	modules[2] = render = new ModuleOpenGL();
-	modules[3] = program = new ModuleProgram("PBR_VertexShader.glsl", "PBR_PixelShader.glsl");
-	modules[4] = camera = new ModuleCamera();
-	modules[5] = fileSystem = new ModuleFileSystem();
-	modules[6] = skybox = new ModuleSkybox();
-	modules[7] = debugDraw = new ModuleDebugDraw();
-	modules[8] = scene = new ModuleScene();
-	modules[9] = editor = new ModuleEditor();
-	modules[10] = clock = new ModuleTimer();
+	modules[2] = camera = new ModuleCamera();
+	modules[3] = fileSystem = new ModuleFileSystem();
+	modules[4] = clock = new ModuleTimer();
+	modules[5] = render = new ModuleOpenGL();
+	modules[6] = debugDraw = new ModuleDebugDraw();
+	modules[7] = scene = new ModuleScene();
+	modules[8] = editor = new ModuleEditor();
 }
 
 Application::~Application()
