+++ Engine/Source/ModuleScene.cpp
@@ -4,7 +4,6 @@
 #include "Application.h"
 #include "ModuleCamera.h"
 #include "glew.h"
-#include "ModuleProgram.h"
 #include "MeshRendererComponent.h"
 #include "Component.h"
