--- Engine/Source/ModuleSkybox.cpp
@@ -1,212 +1,0 @@
-#include "ModuleSkybox.h"
-#include "Application.h"
-#include "ModuleCamera.h"
-#include "ModuleOpenGL.h"
-#include "glew.h"
-
-ModuleSkybox::ModuleSkybox()
-{
-}
-
-ModuleSkybox::~ModuleSkybox()
-{
-}
-
-bool ModuleSkybox::Init()
-{
-    LOG("Intitialize Skybox\n");
-
-    LoadCubeMap();
-
-    float skyboxVertices[] = {
-       -1.0f,  1.0f, -1.0f,
-       -1.0f, -1.0f, -1.0f,
-        1.0f, -1.0f, -1.0f,
-        1.0f, -1.0f, -1.0f,
-        1.0f,  1.0f, -1.0f,
-       -1.0f,  1.0f, -1.0f,
-
-       -1.0f, -1.0f,  1.0f,
-       -1.0f, -1.0f, -1.0f,
-       -1.0f,  1.0f, -1.0f,
-       -1.0f,  1.0f, -1.0f,
-       -1.0f,  1.0f,  1.0f,
-       -1.0f, -1.0f,  1.0f,
-
-        1.0f, -1.0f, -1.0f,
-        1.0f, -1.0f,  1.0f,
-        1.0f,  1.0f,  1.0f,
-        1.0f,  1.0f,  1.0f,
-        1.0f,  1.0f, -1.0f,
-        1.0f, -1.0f, -1.0f,
-
-       -1.0f, -1.0f,  1.0f,
-       -1.0f,  1.0f,  1.0f,
-        1.0f,  1.0f,  1.0f,
-        1.0f,  1.0f,  1.0f,
-        1.0f, -1.0f,  1.0f,
-       -1.0f, -1.0f,  1.0f,
-
-       -1.0f,  1.0f, -1.0f,
-        1.0f,  1.0f, -1.0f,
-        1.0f,  1.0f,  1.0f,
-        1.0f,  1.0f,  1.0f,
-       -1.0f,  1.0f,  1.0f,
-       -1.0f,  1.0f, -1.0f,
-
-       -1.0f, -1.0f, -1.0f,
-       -1.0f, -1.0f,  1.0f,
-        1.0f, -1.0f, -1.0f,
-        1.0f, -1.0f, -1.0f,
-       -1.0f, -1.0f,  1.0f,
-        1.0f, -1.0f,  1.0f
-    };
-
-    glGenVertexArrays(1, &vao);
-    glGenBuffers(1, &vbo);
-    glBindVertexArray(vao);
-    glBindBuffer(GL_ARRAY_BUFFER, vbo);
-    glBufferData(GL_ARRAY_BUFFER, sizeof(skyboxVertices), skyboxVertices, GL_STATIC_DRAW);
-    glEnableVertexAttribArray(0);
-    glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, 0, (void*)0);
-
-    unsigned vtx_shader = CompileShader(GL_VERTEX_SHADER, ReadShader("Assets/Shaders/skybox.vs"));
-    unsigned frg_shader = CompileShader(GL_FRAGMENT_SHADER, ReadShader("Assets/Shaders/skybox.fs"));
-    program = CreateProgram(vtx_shader, frg_shader);
-
-    glActiveTexture(GL_TEXTURE0);
-    glBindTexture(GL_TEXTURE_CUBE_MAP, texture);
-
-    glBindVertexArray(0);
-    return true;
-}
-
-update_status ModuleSkybox::PreUpdate()
-{
-    App->GetOpenGL()->BindSceneFramebuffer();
-
-    proj = App->GetCamera()->GetProjectionMatrix();
-    view = App->GetCamera()->GetViewMatrix();
-
-    glUseProgram(program);
-
-    glBindVertexArray(vao);
-
-    glUniformMatrix4fv(1, 1, GL_TRUE, &view[0][0]);
-    glUniformMatrix4fv(2, 1, GL_TRUE, &proj[0][0]);
-
-    glDepthMask(GL_FALSE);
-
-    glDrawArrays(GL_TRIANGLES, 0, 36);
-    glDepthMask(GL_TRUE);
-
-    glBindVertexArray(0);
-    App->GetOpenGL()->UnbindSceneFramebuffer();
-
-    return update_status::UPDATE_CONTINUE;
-}
-
-bool ModuleSkybox::CleanUp()
-{
-    return true;
-}
-
-unsigned ModuleSkybox::LoadCubeMap()
-{
-    DirectX::ScratchImage image;
-
-    HRESULT res = DirectX::LoadFromDDSFile(L"Assets/Textures/cubemap.dds", DirectX::DDS_FLAGS_NONE, nullptr, image);
-
-    if (res == S_OK)
-    {
-        const DirectX::TexMetadata& metadata = image.GetMetadata();
-
-        glGenTextures(1, &texture);
-        glBindTexture(GL_TEXTURE_CUBE_MAP, texture);
-
-        for (uint32_t i = 0; i < metadata.arraySize; ++i)
-        {
-            const DirectX::Image* face = image.GetImage(0, i, 0);
-
-            glTexImage2D(GL_TEXTURE_CUBE_MAP_POSITIVE_X + i, 0, GL_RGBA8, face->width, face->height, 0, GL_RGBA, GL_UNSIGNED_BYTE, face->pixels);
-        }
-
-        glTexParameteri(GL_TEXTURE_CUBE_MAP, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
-        glTexParameteri(GL_TEXTURE_CUBE_MAP, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
-
-        return texture;
-    }
-
-}
-
-char* ModuleSkybox::ReadShader(const char* filename)
-{
-    char* data = nullptr;
-    FILE* file = nullptr;
-    fopen_s(&file, filename, "rb");
-
-    if (file)
-    {
-        fseek(file, 0, SEEK_END);
-        int size = ftell(file);
-        data = (char*)malloc(size + 1);
-        fseek(file, 0, SEEK_SET);
-        fread(data, 1, size, file);
-        data[size] = 0;
-
-        fclose(file);
-    }
-
-    return data;
-}
-
-unsigned ModuleSkybox::CompileShader(unsigned shaderType, const char* source)
-{
-
-    unsigned shader_id = glCreateShader(shaderType);
-    glShaderSource(shader_id, 1, &source, 0);
-    glCompileShader(shader_id);
-    int res = GL_FALSE;
-    glGetShaderiv(shader_id, GL_COMPILE_STATUS, &res);
-    if (res == GL_FALSE)
-    {
-        int len = 0;
-        glGetShaderiv(shader_id, GL_INFO_LOG_LENGTH, &len);
-        if (len > 0)
-        {
-            int written = 0;
-            char* info = (char*)malloc(len);
-            glGetShaderInfoLog(shader_id, len, &written, info);
-            LOG("Log Info: %s", info);
-            free(info);
-        }
-    }
-    return shader_id;
-
-}
-
-unsigned ModuleSkybox::CreateProgram(unsigned vtx_shader, unsigned frg_shader)
-{
-    unsigned program_id = glCreateProgram();
-    glAttachShader(program_id, vtx_shader);
-    glAttachShader(program_id, frg_shader);
-    glLinkProgram(program_id);
-    int res;
-    glGetProgramiv(program_id, GL_LINK_STATUS, &res);
-    if (res == GL_FALSE)
-    {
-        int len = 0;
-        glGetProgramiv(program_id, GL_INFO_LOG_LENGTH, &len);
-        if (len > 0)
-        {
-            int written = 0;
-            char* info = (char*)malloc(len);
-            glGetProgramInfoLog(program_id, len, &written, info);
-            LOG("Program Log Info: %s", info);
-            free(info);
-        }
-    }
-    glDeleteShader(vtx_shader);
-    glDeleteShader(frg_shader);
-    return program_id;
-}
